% Generated by roxygen2 (4.0.1): do not edit by hand
\name{declare}
\alias{declare}
\title{Declare symbols}
\usage{
declare(..., .envir = parent.frame())
}
\arguments{
\item{...}{Symbols to declare}

\item{.envir}{\code{environment} to store the symbols}
}
\description{
Declare symbols
}
\examples{
declare(x,y=numeric(),z=integer())
}

